third_party/blink/web_tests/platform/mac/external/wpt/css/css-shapes/shape-outside/values/shape-outside-ellipse-004-expected.txt

This is a testharness.js-based test.
Found 780 tests; 468 PASS, 312 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] test unit (inline): cm - ellipse(at 50cm)
[PASS] test unit (inline): cm - ellipse(at 50% 50cm)
[PASS] test unit (inline): cm - ellipse(at 50cm 50%)
[PASS] test unit (inline): cm - ellipse(at 50cm 50cm)
[PASS] test unit (inline): cm - ellipse(at left 50cm)
[PASS] test unit (inline): cm - ellipse(at 50cm top)
[PASS] test unit (inline): cm - ellipse(at right 80cm)
[PASS] test unit (inline): cm - ellipse(at 70cm bottom)
[PASS] test unit (inline): cm - ellipse(at center 60cm)
[PASS] test unit (inline): cm - ellipse(at 60cm center)
[PASS] test unit (inline): cm - ellipse(at left 50% top 50cm)
[FAIL] test unit (inline): cm - ellipse(at left 50% bottom 70cm)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 70cm)"
[PASS] test unit (inline): cm - ellipse(at left 50cm top 50%)
[PASS] test unit (inline): cm - ellipse(at left 50cm top 50cm)
[FAIL] test unit (inline): cm - ellipse(at left 50cm bottom 70%)
  assert_equals: expected "ellipse(at 50cm calc(30%))" but got "ellipse(at 50cm 30%)"
[PASS] test unit (inline): cm - ellipse(at top 50% left 50cm)
[FAIL] test unit (inline): cm - ellipse(at top 50% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 80cm top 50%)"
[PASS] test unit (inline): cm - ellipse(at top 50cm left 50%)
[PASS] test unit (inline): cm - ellipse(at top 50cm left 50cm)
[FAIL] test unit (inline): cm - ellipse(at top 50cm right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50cm)" but got "ellipse(at 20% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70% left 50cm)
  assert_equals: expected "ellipse(at 50cm calc(30%))" but got "ellipse(at 50cm 30%)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(30%))" but got "ellipse(at right 80cm top 30%)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70cm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70cm right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70cm right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 80cm bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80% top 50cm)
  assert_equals: expected "ellipse(at calc(20%) 50cm)" but got "ellipse(at 20% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80% bottom 70cm)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 2645.67px))" but got "ellipse(at left 20% bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80cm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 80cm top 50%)"
[FAIL] test unit (inline): cm - ellipse(at right 80cm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(30%))" but got "ellipse(at right 80cm top 30%)"
[FAIL] test unit (inline): cm - ellipse(at right 80cm bottom 70cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 80cm bottom 70cm)"
[PASS] test unit (inline): mm - ellipse(at 50mm)
[PASS] test unit (inline): mm - ellipse(at 50% 50mm)
[PASS] test unit (inline): mm - ellipse(at 50mm 50%)
[PASS] test unit (inline): mm - ellipse(at 50mm 50mm)
[PASS] test unit (inline): mm - ellipse(at left 50mm)
[PASS] test unit (inline): mm - ellipse(at 50mm top)
[PASS] test unit (inline): mm - ellipse(at right 80mm)
[PASS] test unit (inline): mm - ellipse(at 70mm bottom)
[PASS] test unit (inline): mm - ellipse(at center 60mm)
[PASS] test unit (inline): mm - ellipse(at 60mm center)
[PASS] test unit (inline): mm - ellipse(at left 50% top 50mm)
[FAIL] test unit (inline): mm - ellipse(at left 50% bottom 70mm)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 70mm)"
[PASS] test unit (inline): mm - ellipse(at left 50mm top 50%)
[PASS] test unit (inline): mm - ellipse(at left 50mm top 50mm)
[FAIL] test unit (inline): mm - ellipse(at left 50mm bottom 70%)
  assert_equals: expected "ellipse(at 50mm calc(30%))" but got "ellipse(at 50mm 30%)"
[PASS] test unit (inline): mm - ellipse(at top 50% left 50mm)
[FAIL] test unit (inline): mm - ellipse(at top 50% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 80mm top 50%)"
[PASS] test unit (inline): mm - ellipse(at top 50mm left 50%)
[PASS] test unit (inline): mm - ellipse(at top 50mm left 50mm)
[FAIL] test unit (inline): mm - ellipse(at top 50mm right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50mm)" but got "ellipse(at 20% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70% left 50mm)
  assert_equals: expected "ellipse(at 50mm calc(30%))" but got "ellipse(at 50mm 30%)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(30%))" but got "ellipse(at right 80mm top 30%)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70mm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70mm right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70mm right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 80mm bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80% top 50mm)
  assert_equals: expected "ellipse(at calc(20%) 50mm)" but got "ellipse(at 20% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80% bottom 70mm)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 264.567px))" but got "ellipse(at left 20% bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80mm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 80mm top 50%)"
[FAIL] test unit (inline): mm - ellipse(at right 80mm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(30%))" but got "ellipse(at right 80mm top 30%)"
[FAIL] test unit (inline): mm - ellipse(at right 80mm bottom 70mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 80mm bottom 70mm)"
[PASS] test unit (inline): in - ellipse(at 50in)
[PASS] test unit (inline): in - ellipse(at 50% 50in)
[PASS] test unit (inline): in - ellipse(at 50in 50%)
[PASS] test unit (inline): in - ellipse(at 50in 50in)
[PASS] test unit (inline): in - ellipse(at left 50in)
[PASS] test unit (inline): in - ellipse(at 50in top)
[PASS] test unit (inline): in - ellipse(at right 80in)
[PASS] test unit (inline): in - ellipse(at 70in bottom)
[PASS] test unit (inline): in - ellipse(at center 60in)
[PASS] test unit (inline): in - ellipse(at 60in center)
[PASS] test unit (inline): in - ellipse(at left 50% top 50in)
[FAIL] test unit (inline): in - ellipse(at left 50% bottom 70in)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 70in)"
[PASS] test unit (inline): in - ellipse(at left 50in top 50%)
[PASS] test unit (inline): in - ellipse(at left 50in top 50in)
[FAIL] test unit (inline): in - ellipse(at left 50in bottom 70%)
  assert_equals: expected "ellipse(at 50in calc(30%))" but got "ellipse(at 50in 30%)"
[PASS] test unit (inline): in - ellipse(at top 50% left 50in)
[FAIL] test unit (inline): in - ellipse(at top 50% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 80in top 50%)"
[PASS] test unit (inline): in - ellipse(at top 50in left 50%)
[PASS] test unit (inline): in - ellipse(at top 50in left 50in)
[FAIL] test unit (inline): in - ellipse(at top 50in right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50in)" but got "ellipse(at 20% 50in)"
[FAIL] test unit (inline): in - ellipse(at bottom 70% left 50in)
  assert_equals: expected "ellipse(at 50in calc(30%))" but got "ellipse(at 50in 30%)"
[FAIL] test unit (inline): in - ellipse(at bottom 70% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(30%))" but got "ellipse(at right 80in top 30%)"
[FAIL] test unit (inline): in - ellipse(at bottom 70in left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at bottom 70in right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 6720px))" but got "ellipse(at left 50% bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at bottom 70in right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 80in bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at right 80% top 50in)
  assert_equals: expected "ellipse(at calc(20%) 50in)" but got "ellipse(at 20% 50in)"
[FAIL] test unit (inline): in - ellipse(at right 80% bottom 70in)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 6720px))" but got "ellipse(at left 20% bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at right 80in top 50%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 80in top 50%)"
[FAIL] test unit (inline): in - ellipse(at right 80in bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(30%))" but got "ellipse(at right 80in top 30%)"
[FAIL] test unit (inline): in - ellipse(at right 80in bottom 70in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 80in bottom 70in)"
[PASS] test unit (inline): pt - ellipse(at 50pt)
[PASS] test unit (inline): pt - ellipse(at 50% 50pt)
[PASS] test unit (inline): pt - ellipse(at 50pt 50%)
[PASS] test unit (inline): pt - ellipse(at 50pt 50pt)
[PASS] test unit (inline): pt - ellipse(at left 50pt)
[PASS] test unit (inline): pt - ellipse(at 50pt top)
[PASS] test unit (inline): pt - ellipse(at right 80pt)
[PASS] test unit (inline): pt - ellipse(at 70pt bottom)
[PASS] test unit (inline): pt - ellipse(at center 60pt)
[PASS] test unit (inline): pt - ellipse(at 60pt center)
[PASS] test unit (inline): pt - ellipse(at left 50% top 50pt)
[FAIL] test unit (inline): pt - ellipse(at left 50% bottom 70pt)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 70pt)"
[PASS] test unit (inline): pt - ellipse(at left 50pt top 50%)
[PASS] test unit (inline): pt - ellipse(at left 50pt top 50pt)
[FAIL] test unit (inline): pt - ellipse(at left 50pt bottom 70%)
  assert_equals: expected "ellipse(at 50pt calc(30%))" but got "ellipse(at 50pt 30%)"
[PASS] test unit (inline): pt - ellipse(at top 50% left 50pt)
[FAIL] test unit (inline): pt - ellipse(at top 50% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 80pt top 50%)"
[PASS] test unit (inline): pt - ellipse(at top 50pt left 50%)
[PASS] test unit (inline): pt - ellipse(at top 50pt left 50pt)
[FAIL] test unit (inline): pt - ellipse(at top 50pt right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50pt)" but got "ellipse(at 20% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70% left 50pt)
  assert_equals: expected "ellipse(at 50pt calc(30%))" but got "ellipse(at 50pt 30%)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(30%))" but got "ellipse(at right 80pt top 30%)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70pt left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70pt right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70pt right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 80pt bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80% top 50pt)
  assert_equals: expected "ellipse(at calc(20%) 50pt)" but got "ellipse(at 20% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80% bottom 70pt)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 93.3333px))" but got "ellipse(at left 20% bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80pt top 50%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 80pt top 50%)"
[FAIL] test unit (inline): pt - ellipse(at right 80pt bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(30%))" but got "ellipse(at right 80pt top 30%)"
[FAIL] test unit (inline): pt - ellipse(at right 80pt bottom 70pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 80pt bottom 70pt)"
[PASS] test unit (inline): pc - ellipse(at 50pc)
[PASS] test unit (inline): pc - ellipse(at 50% 50pc)
[PASS] test unit (inline): pc - ellipse(at 50pc 50%)
[PASS] test unit (inline): pc - ellipse(at 50pc 50pc)
[PASS] test unit (inline): pc - ellipse(at left 50pc)
[PASS] test unit (inline): pc - ellipse(at 50pc top)
[PASS] test unit (inline): pc - ellipse(at right 80pc)
[PASS] test unit (inline): pc - ellipse(at 70pc bottom)
[PASS] test unit (inline): pc - ellipse(at center 60pc)
[PASS] test unit (inline): pc - ellipse(at 60pc center)
[PASS] test unit (inline): pc - ellipse(at left 50% top 50pc)
[FAIL] test unit (inline): pc - ellipse(at left 50% bottom 70pc)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 70pc)"
[PASS] test unit (inline): pc - ellipse(at left 50pc top 50%)
[PASS] test unit (inline): pc - ellipse(at left 50pc top 50pc)
[FAIL] test unit (inline): pc - ellipse(at left 50pc bottom 70%)
  assert_equals: expected "ellipse(at 50pc calc(30%))" but got "ellipse(at 50pc 30%)"
[PASS] test unit (inline): pc - ellipse(at top 50% left 50pc)
[FAIL] test unit (inline): pc - ellipse(at top 50% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 80pc top 50%)"
[PASS] test unit (inline): pc - ellipse(at top 50pc left 50%)
[PASS] test unit (inline): pc - ellipse(at top 50pc left 50pc)
[FAIL] test unit (inline): pc - ellipse(at top 50pc right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50pc)" but got "ellipse(at 20% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70% left 50pc)
  assert_equals: expected "ellipse(at 50pc calc(30%))" but got "ellipse(at 50pc 30%)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(30%))" but got "ellipse(at right 80pc top 30%)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70pc left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70pc right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 1120px))" but got "ellipse(at left 50% bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70pc right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 80pc bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80% top 50pc)
  assert_equals: expected "ellipse(at calc(20%) 50pc)" but got "ellipse(at 20% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80% bottom 70pc)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 1120px))" but got "ellipse(at left 20% bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80pc top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 80pc top 50%)"
[FAIL] test unit (inline): pc - ellipse(at right 80pc bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(30%))" but got "ellipse(at right 80pc top 30%)"
[FAIL] test unit (inline): pc - ellipse(at right 80pc bottom 70pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 80pc bottom 70pc)"
[PASS] test unit (inline): em - ellipse(at 50em)
[PASS] test unit (inline): em - ellipse(at 50% 50em)
[PASS] test unit (inline): em - ellipse(at 50em 50%)
[PASS] test unit (inline): em - ellipse(at 50em 50em)
[PASS] test unit (inline): em - ellipse(at left 50em)
[PASS] test unit (inline): em - ellipse(at 50em top)
[PASS] test unit (inline): em - ellipse(at right 80em)
[PASS] test unit (inline): em - ellipse(at 70em bottom)
[PASS] test unit (inline): em - ellipse(at center 60em)
[PASS] test unit (inline): em - ellipse(at 60em center)
[PASS] test unit (inline): em - ellipse(at left 50% top 50em)
[FAIL] test unit (inline): em - ellipse(at left 50% bottom 70em)
  assert_equals: expected "ellipse(at 50% calc(100% - 70em))" but got "ellipse(at left 50% bottom 70em)"
[PASS] test unit (inline): em - ellipse(at left 50em top 50%)
[PASS] test unit (inline): em - ellipse(at left 50em top 50em)
[FAIL] test unit (inline): em - ellipse(at left 50em bottom 70%)
  assert_equals: expected "ellipse(at 50em calc(30%))" but got "ellipse(at 50em 30%)"
[PASS] test unit (inline): em - ellipse(at top 50% left 50em)
[FAIL] test unit (inline): em - ellipse(at top 50% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 80em) 50%)" but got "ellipse(at right 80em top 50%)"
[PASS] test unit (inline): em - ellipse(at top 50em left 50%)
[PASS] test unit (inline): em - ellipse(at top 50em left 50em)
[FAIL] test unit (inline): em - ellipse(at top 50em right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50em)" but got "ellipse(at 20% 50em)"
[FAIL] test unit (inline): em - ellipse(at bottom 70% left 50em)
  assert_equals: expected "ellipse(at 50em calc(30%))" but got "ellipse(at 50em 30%)"
[FAIL] test unit (inline): em - ellipse(at bottom 70% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 80em) calc(30%))" but got "ellipse(at right 80em top 30%)"
[FAIL] test unit (inline): em - ellipse(at bottom 70em left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70em))" but got "ellipse(at left 50% bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at bottom 70em right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70em))" but got "ellipse(at left 50% bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at bottom 70em right 80em)
  assert_equals: expected "ellipse(at calc(100% - 80em) calc(100% - 70em))" but got "ellipse(at right 80em bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at right 80% top 50em)
  assert_equals: expected "ellipse(at calc(20%) 50em)" but got "ellipse(at 20% 50em)"
[FAIL] test unit (inline): em - ellipse(at right 80% bottom 70em)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70em))" but got "ellipse(at left 20% bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at right 80em top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80em) 50%)" but got "ellipse(at right 80em top 50%)"
[FAIL] test unit (inline): em - ellipse(at right 80em bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80em) calc(30%))" but got "ellipse(at right 80em top 30%)"
[FAIL] test unit (inline): em - ellipse(at right 80em bottom 70em)
  assert_equals: expected "ellipse(at calc(100% - 80em) calc(100% - 70em))" but got "ellipse(at right 80em bottom 70em)"
[PASS] test unit (inline): ex - ellipse(at 50ex)
[PASS] test unit (inline): ex - ellipse(at 50% 50ex)
[PASS] test unit (inline): ex - ellipse(at 50ex 50%)
[PASS] test unit (inline): ex - ellipse(at 50ex 50ex)
[PASS] test unit (inline): ex - ellipse(at left 50ex)
[PASS] test unit (inline): ex - ellipse(at 50ex top)
[PASS] test unit (inline): ex - ellipse(at right 80ex)
[PASS] test unit (inline): ex - ellipse(at 70ex bottom)
[PASS] test unit (inline): ex - ellipse(at center 60ex)
[PASS] test unit (inline): ex - ellipse(at 60ex center)
[PASS] test unit (inline): ex - ellipse(at left 50% top 50ex)
[FAIL] test unit (inline): ex - ellipse(at left 50% bottom 70ex)
  assert_equals: expected "ellipse(at 50% calc(100% - 70ex))" but got "ellipse(at left 50% bottom 70ex)"
[PASS] test unit (inline): ex - ellipse(at left 50ex top 50%)
[PASS] test unit (inline): ex - ellipse(at left 50ex top 50ex)
[FAIL] test unit (inline): ex - ellipse(at left 50ex bottom 70%)
  assert_equals: expected "ellipse(at 50ex calc(30%))" but got "ellipse(at 50ex 30%)"
[PASS] test unit (inline): ex - ellipse(at top 50% left 50ex)
[FAIL] test unit (inline): ex - ellipse(at top 50% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 80ex) 50%)" but got "ellipse(at right 80ex top 50%)"
[PASS] test unit (inline): ex - ellipse(at top 50ex left 50%)
[PASS] test unit (inline): ex - ellipse(at top 50ex left 50ex)
[FAIL] test unit (inline): ex - ellipse(at top 50ex right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50ex)" but got "ellipse(at 20% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70% left 50ex)
  assert_equals: expected "ellipse(at 50ex calc(30%))" but got "ellipse(at 50ex 30%)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 80ex) calc(30%))" but got "ellipse(at right 80ex top 30%)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70ex left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70ex))" but got "ellipse(at left 50% bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70ex right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70ex))" but got "ellipse(at left 50% bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70ex right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 80ex) calc(100% - 70ex))" but got "ellipse(at right 80ex bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80% top 50ex)
  assert_equals: expected "ellipse(at calc(20%) 50ex)" but got "ellipse(at 20% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80% bottom 70ex)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70ex))" but got "ellipse(at left 20% bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80ex top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80ex) 50%)" but got "ellipse(at right 80ex top 50%)"
[FAIL] test unit (inline): ex - ellipse(at right 80ex bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80ex) calc(30%))" but got "ellipse(at right 80ex top 30%)"
[FAIL] test unit (inline): ex - ellipse(at right 80ex bottom 70ex)
  assert_equals: expected "ellipse(at calc(100% - 80ex) calc(100% - 70ex))" but got "ellipse(at right 80ex bottom 70ex)"
[PASS] test unit (inline): ch - ellipse(at 50ch)
[PASS] test unit (inline): ch - ellipse(at 50% 50ch)
[PASS] test unit (inline): ch - ellipse(at 50ch 50%)
[PASS] test unit (inline): ch - ellipse(at 50ch 50ch)
[PASS] test unit (inline): ch - ellipse(at left 50ch)
[PASS] test unit (inline): ch - ellipse(at 50ch top)
[PASS] test unit (inline): ch - ellipse(at right 80ch)
[PASS] test unit (inline): ch - ellipse(at 70ch bottom)
[PASS] test unit (inline): ch - ellipse(at center 60ch)
[PASS] test unit (inline): ch - ellipse(at 60ch center)
[PASS] test unit (inline): ch - ellipse(at left 50% top 50ch)
[FAIL] test unit (inline): ch - ellipse(at left 50% bottom 70ch)
  assert_equals: expected "ellipse(at 50% calc(100% - 70ch))" but got "ellipse(at left 50% bottom 70ch)"
[PASS] test unit (inline): ch - ellipse(at left 50ch top 50%)
[PASS] test unit (inline): ch - ellipse(at left 50ch top 50ch)
[FAIL] test unit (inline): ch - ellipse(at left 50ch bottom 70%)
  assert_equals: expected "ellipse(at 50ch calc(30%))" but got "ellipse(at 50ch 30%)"
[PASS] test unit (inline): ch - ellipse(at top 50% left 50ch)
[FAIL] test unit (inline): ch - ellipse(at top 50% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 80ch) 50%)" but got "ellipse(at right 80ch top 50%)"
[PASS] test unit (inline): ch - ellipse(at top 50ch left 50%)
[PASS] test unit (inline): ch - ellipse(at top 50ch left 50ch)
[FAIL] test unit (inline): ch - ellipse(at top 50ch right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50ch)" but got "ellipse(at 20% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70% left 50ch)
  assert_equals: expected "ellipse(at 50ch calc(30%))" but got "ellipse(at 50ch 30%)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 80ch) calc(30%))" but got "ellipse(at right 80ch top 30%)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70ch left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70ch))" but got "ellipse(at left 50% bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70ch right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70ch))" but got "ellipse(at left 50% bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70ch right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 80ch) calc(100% - 70ch))" but got "ellipse(at right 80ch bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80% top 50ch)
  assert_equals: expected "ellipse(at calc(20%) 50ch)" but got "ellipse(at 20% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80% bottom 70ch)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70ch))" but got "ellipse(at left 20% bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80ch top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80ch) 50%)" but got "ellipse(at right 80ch top 50%)"
[FAIL] test unit (inline): ch - ellipse(at right 80ch bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80ch) calc(30%))" but got "ellipse(at right 80ch top 30%)"
[FAIL] test unit (inline): ch - ellipse(at right 80ch bottom 70ch)
  assert_equals: expected "ellipse(at calc(100% - 80ch) calc(100% - 70ch))" but got "ellipse(at right 80ch bottom 70ch)"
[PASS] test unit (inline): rem - ellipse(at 50rem)
[PASS] test unit (inline): rem - ellipse(at 50% 50rem)
[PASS] test unit (inline): rem - ellipse(at 50rem 50%)
[PASS] test unit (inline): rem - ellipse(at 50rem 50rem)
[PASS] test unit (inline): rem - ellipse(at left 50rem)
[PASS] test unit (inline): rem - ellipse(at 50rem top)
[PASS] test unit (inline): rem - ellipse(at right 80rem)
[PASS] test unit (inline): rem - ellipse(at 70rem bottom)
[PASS] test unit (inline): rem - ellipse(at center 60rem)
[PASS] test unit (inline): rem - ellipse(at 60rem center)
[PASS] test unit (inline): rem - ellipse(at left 50% top 50rem)
[FAIL] test unit (inline): rem - ellipse(at left 50% bottom 70rem)
  assert_equals: expected "ellipse(at 50% calc(100% - 70rem))" but got "ellipse(at left 50% bottom 70rem)"
[PASS] test unit (inline): rem - ellipse(at left 50rem top 50%)
[PASS] test unit (inline): rem - ellipse(at left 50rem top 50rem)
[FAIL] test unit (inline): rem - ellipse(at left 50rem bottom 70%)
  assert_equals: expected "ellipse(at 50rem calc(30%))" but got "ellipse(at 50rem 30%)"
[PASS] test unit (inline): rem - ellipse(at top 50% left 50rem)
[FAIL] test unit (inline): rem - ellipse(at top 50% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 80rem) 50%)" but got "ellipse(at right 80rem top 50%)"
[PASS] test unit (inline): rem - ellipse(at top 50rem left 50%)
[PASS] test unit (inline): rem - ellipse(at top 50rem left 50rem)
[FAIL] test unit (inline): rem - ellipse(at top 50rem right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50rem)" but got "ellipse(at 20% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70% left 50rem)
  assert_equals: expected "ellipse(at 50rem calc(30%))" but got "ellipse(at 50rem 30%)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 80rem) calc(30%))" but got "ellipse(at right 80rem top 30%)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70rem left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70rem))" but got "ellipse(at left 50% bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70rem right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70rem))" but got "ellipse(at left 50% bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70rem right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 80rem) calc(100% - 70rem))" but got "ellipse(at right 80rem bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80% top 50rem)
  assert_equals: expected "ellipse(at calc(20%) 50rem)" but got "ellipse(at 20% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80% bottom 70rem)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70rem))" but got "ellipse(at left 20% bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80rem top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80rem) 50%)" but got "ellipse(at right 80rem top 50%)"
[FAIL] test unit (inline): rem - ellipse(at right 80rem bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80rem) calc(30%))" but got "ellipse(at right 80rem top 30%)"
[FAIL] test unit (inline): rem - ellipse(at right 80rem bottom 70rem)
  assert_equals: expected "ellipse(at calc(100% - 80rem) calc(100% - 70rem))" but got "ellipse(at right 80rem bottom 70rem)"
[PASS] test unit (inline): vw - ellipse(at 50vw)
[PASS] test unit (inline): vw - ellipse(at 50% 50vw)
[PASS] test unit (inline): vw - ellipse(at 50vw 50%)
[PASS] test unit (inline): vw - ellipse(at 50vw 50vw)
[PASS] test unit (inline): vw - ellipse(at left 50vw)
[PASS] test unit (inline): vw - ellipse(at 50vw top)
[PASS] test unit (inline): vw - ellipse(at right 80vw)
[PASS] test unit (inline): vw - ellipse(at 70vw bottom)
[PASS] test unit (inline): vw - ellipse(at center 60vw)
[PASS] test unit (inline): vw - ellipse(at 60vw center)
[PASS] test unit (inline): vw - ellipse(at left 50% top 50vw)
[FAIL] test unit (inline): vw - ellipse(at left 50% bottom 70vw)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vw))" but got "ellipse(at left 50% bottom 70vw)"
[PASS] test unit (inline): vw - ellipse(at left 50vw top 50%)
[PASS] test unit (inline): vw - ellipse(at left 50vw top 50vw)
[FAIL] test unit (inline): vw - ellipse(at left 50vw bottom 70%)
  assert_equals: expected "ellipse(at 50vw calc(30%))" but got "ellipse(at 50vw 30%)"
[PASS] test unit (inline): vw - ellipse(at top 50% left 50vw)
[FAIL] test unit (inline): vw - ellipse(at top 50% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 80vw) 50%)" but got "ellipse(at right 80vw top 50%)"
[PASS] test unit (inline): vw - ellipse(at top 50vw left 50%)
[PASS] test unit (inline): vw - ellipse(at top 50vw left 50vw)
[FAIL] test unit (inline): vw - ellipse(at top 50vw right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50vw)" but got "ellipse(at 20% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70% left 50vw)
  assert_equals: expected "ellipse(at 50vw calc(30%))" but got "ellipse(at 50vw 30%)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 80vw) calc(30%))" but got "ellipse(at right 80vw top 30%)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70vw left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vw))" but got "ellipse(at left 50% bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70vw right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70vw))" but got "ellipse(at left 50% bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70vw right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 80vw) calc(100% - 70vw))" but got "ellipse(at right 80vw bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80% top 50vw)
  assert_equals: expected "ellipse(at calc(20%) 50vw)" but got "ellipse(at 20% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80% bottom 70vw)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70vw))" but got "ellipse(at left 20% bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80vw top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80vw) 50%)" but got "ellipse(at right 80vw top 50%)"
[FAIL] test unit (inline): vw - ellipse(at right 80vw bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80vw) calc(30%))" but got "ellipse(at right 80vw top 30%)"
[FAIL] test unit (inline): vw - ellipse(at right 80vw bottom 70vw)
  assert_equals: expected "ellipse(at calc(100% - 80vw) calc(100% - 70vw))" but got "ellipse(at right 80vw bottom 70vw)"
[PASS] test unit (inline): vh - ellipse(at 50vh)
[PASS] test unit (inline): vh - ellipse(at 50% 50vh)
[PASS] test unit (inline): vh - ellipse(at 50vh 50%)
[PASS] test unit (inline): vh - ellipse(at 50vh 50vh)
[PASS] test unit (inline): vh - ellipse(at left 50vh)
[PASS] test unit (inline): vh - ellipse(at 50vh top)
[PASS] test unit (inline): vh - ellipse(at right 80vh)
[PASS] test unit (inline): vh - ellipse(at 70vh bottom)
[PASS] test unit (inline): vh - ellipse(at center 60vh)
[PASS] test unit (inline): vh - ellipse(at 60vh center)
[PASS] test unit (inline): vh - ellipse(at left 50% top 50vh)
[FAIL] test unit (inline): vh - ellipse(at left 50% bottom 70vh)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vh))" but got "ellipse(at left 50% bottom 70vh)"
[PASS] test unit (inline): vh - ellipse(at left 50vh top 50%)
[PASS] test unit (inline): vh - ellipse(at left 50vh top 50vh)
[FAIL] test unit (inline): vh - ellipse(at left 50vh bottom 70%)
  assert_equals: expected "ellipse(at 50vh calc(30%))" but got "ellipse(at 50vh 30%)"
[PASS] test unit (inline): vh - ellipse(at top 50% left 50vh)
[FAIL] test unit (inline): vh - ellipse(at top 50% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 80vh) 50%)" but got "ellipse(at right 80vh top 50%)"
[PASS] test unit (inline): vh - ellipse(at top 50vh left 50%)
[PASS] test unit (inline): vh - ellipse(at top 50vh left 50vh)
[FAIL] test unit (inline): vh - ellipse(at top 50vh right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50vh)" but got "ellipse(at 20% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70% left 50vh)
  assert_equals: expected "ellipse(at 50vh calc(30%))" but got "ellipse(at 50vh 30%)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 80vh) calc(30%))" but got "ellipse(at right 80vh top 30%)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70vh left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vh))" but got "ellipse(at left 50% bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70vh right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70vh))" but got "ellipse(at left 50% bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70vh right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 80vh) calc(100% - 70vh))" but got "ellipse(at right 80vh bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80% top 50vh)
  assert_equals: expected "ellipse(at calc(20%) 50vh)" but got "ellipse(at 20% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80% bottom 70vh)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70vh))" but got "ellipse(at left 20% bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80vh top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80vh) 50%)" but got "ellipse(at right 80vh top 50%)"
[FAIL] test unit (inline): vh - ellipse(at right 80vh bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80vh) calc(30%))" but got "ellipse(at right 80vh top 30%)"
[FAIL] test unit (inline): vh - ellipse(at right 80vh bottom 70vh)
  assert_equals: expected "ellipse(at calc(100% - 80vh) calc(100% - 70vh))" but got "ellipse(at right 80vh bottom 70vh)"
[PASS] test unit (inline): vmin - ellipse(at 50vmin)
[PASS] test unit (inline): vmin - ellipse(at 50% 50vmin)
[PASS] test unit (inline): vmin - ellipse(at 50vmin 50%)
[PASS] test unit (inline): vmin - ellipse(at 50vmin 50vmin)
[PASS] test unit (inline): vmin - ellipse(at left 50vmin)
[PASS] test unit (inline): vmin - ellipse(at 50vmin top)
[PASS] test unit (inline): vmin - ellipse(at right 80vmin)
[PASS] test unit (inline): vmin - ellipse(at 70vmin bottom)
[PASS] test unit (inline): vmin - ellipse(at center 60vmin)
[PASS] test unit (inline): vmin - ellipse(at 60vmin center)
[PASS] test unit (inline): vmin - ellipse(at left 50% top 50vmin)
[FAIL] test unit (inline): vmin - ellipse(at left 50% bottom 70vmin)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vmin))" but got "ellipse(at left 50% bottom 70vmin)"
[PASS] test unit (inline): vmin - ellipse(at left 50vmin top 50%)
[PASS] test unit (inline): vmin - ellipse(at left 50vmin top 50vmin)
[FAIL] test unit (inline): vmin - ellipse(at left 50vmin bottom 70%)
  assert_equals: expected "ellipse(at 50vmin calc(30%))" but got "ellipse(at 50vmin 30%)"
[PASS] test unit (inline): vmin - ellipse(at top 50% left 50vmin)
[FAIL] test unit (inline): vmin - ellipse(at top 50% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) 50%)" but got "ellipse(at right 80vmin top 50%)"
[PASS] test unit (inline): vmin - ellipse(at top 50vmin left 50%)
[PASS] test unit (inline): vmin - ellipse(at top 50vmin left 50vmin)
[FAIL] test unit (inline): vmin - ellipse(at top 50vmin right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50vmin)" but got "ellipse(at 20% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70% left 50vmin)
  assert_equals: expected "ellipse(at 50vmin calc(30%))" but got "ellipse(at 50vmin 30%)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) calc(30%))" but got "ellipse(at right 80vmin top 30%)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70vmin left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vmin))" but got "ellipse(at left 50% bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70vmin right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70vmin))" but got "ellipse(at left 50% bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70vmin right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) calc(100% - 70vmin))" but got "ellipse(at right 80vmin bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80% top 50vmin)
  assert_equals: expected "ellipse(at calc(20%) 50vmin)" but got "ellipse(at 20% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80% bottom 70vmin)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70vmin))" but got "ellipse(at left 20% bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80vmin top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) 50%)" but got "ellipse(at right 80vmin top 50%)"
[FAIL] test unit (inline): vmin - ellipse(at right 80vmin bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) calc(30%))" but got "ellipse(at right 80vmin top 30%)"
[FAIL] test unit (inline): vmin - ellipse(at right 80vmin bottom 70vmin)
  assert_equals: expected "ellipse(at calc(100% - 80vmin) calc(100% - 70vmin))" but got "ellipse(at right 80vmin bottom 70vmin)"
[PASS] test unit (inline): vmax - ellipse(at 50vmax)
[PASS] test unit (inline): vmax - ellipse(at 50% 50vmax)
[PASS] test unit (inline): vmax - ellipse(at 50vmax 50%)
[PASS] test unit (inline): vmax - ellipse(at 50vmax 50vmax)
[PASS] test unit (inline): vmax - ellipse(at left 50vmax)
[PASS] test unit (inline): vmax - ellipse(at 50vmax top)
[PASS] test unit (inline): vmax - ellipse(at right 80vmax)
[PASS] test unit (inline): vmax - ellipse(at 70vmax bottom)
[PASS] test unit (inline): vmax - ellipse(at center 60vmax)
[PASS] test unit (inline): vmax - ellipse(at 60vmax center)
[PASS] test unit (inline): vmax - ellipse(at left 50% top 50vmax)
[FAIL] test unit (inline): vmax - ellipse(at left 50% bottom 70vmax)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vmax))" but got "ellipse(at left 50% bottom 70vmax)"
[PASS] test unit (inline): vmax - ellipse(at left 50vmax top 50%)
[PASS] test unit (inline): vmax - ellipse(at left 50vmax top 50vmax)
[FAIL] test unit (inline): vmax - ellipse(at left 50vmax bottom 70%)
  assert_equals: expected "ellipse(at 50vmax calc(30%))" but got "ellipse(at 50vmax 30%)"
[PASS] test unit (inline): vmax - ellipse(at top 50% left 50vmax)
[FAIL] test unit (inline): vmax - ellipse(at top 50% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) 50%)" but got "ellipse(at right 80vmax top 50%)"
[PASS] test unit (inline): vmax - ellipse(at top 50vmax left 50%)
[PASS] test unit (inline): vmax - ellipse(at top 50vmax left 50vmax)
[FAIL] test unit (inline): vmax - ellipse(at top 50vmax right 80%)
  assert_equals: expected "ellipse(at calc(20%) 50vmax)" but got "ellipse(at 20% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70% left 50vmax)
  assert_equals: expected "ellipse(at 50vmax calc(30%))" but got "ellipse(at 50vmax 30%)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) calc(30%))" but got "ellipse(at right 80vmax top 30%)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70vmax left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 70vmax))" but got "ellipse(at left 50% bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70vmax right 50%)
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70vmax))" but got "ellipse(at left 50% bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70vmax right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) calc(100% - 70vmax))" but got "ellipse(at right 80vmax bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80% top 50vmax)
  assert_equals: expected "ellipse(at calc(20%) 50vmax)" but got "ellipse(at 20% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80% bottom 70vmax)
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70vmax))" but got "ellipse(at left 20% bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80vmax top 50%)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) 50%)" but got "ellipse(at right 80vmax top 50%)"
[FAIL] test unit (inline): vmax - ellipse(at right 80vmax bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) calc(30%))" but got "ellipse(at right 80vmax top 30%)"
[FAIL] test unit (inline): vmax - ellipse(at right 80vmax bottom 70vmax)
  assert_equals: expected "ellipse(at calc(100% - 80vmax) calc(100% - 70vmax))" but got "ellipse(at right 80vmax bottom 70vmax)"
[PASS] test unit (computed): cm - ellipse(at 50cm)
[PASS] test unit (computed): cm - ellipse(at 50% 50cm)
[PASS] test unit (computed): cm - ellipse(at 50cm 50%)
[PASS] test unit (computed): cm - ellipse(at 50cm 50cm)
[PASS] test unit (computed): cm - ellipse(at left 50cm)
[PASS] test unit (computed): cm - ellipse(at 50cm top)
[PASS] test unit (computed): cm - ellipse(at right 80cm)
[PASS] test unit (computed): cm - ellipse(at 70cm bottom)
[PASS] test unit (computed): cm - ellipse(at center 60cm)
[PASS] test unit (computed): cm - ellipse(at 60cm center)
[PASS] test unit (computed): cm - ellipse(at left 50% top 50cm)
[FAIL] test unit (computed): cm - ellipse(at left 50% bottom 70cm)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[PASS] test unit (computed): cm - ellipse(at left 50cm top 50%)
[PASS] test unit (computed): cm - ellipse(at left 50cm top 50cm)
[PASS] test unit (computed): cm - ellipse(at left 50cm bottom 70%)
[PASS] test unit (computed): cm - ellipse(at top 50% left 50cm)
[FAIL] test unit (computed): cm - ellipse(at top 50% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 3023.62px top 50%)"
[PASS] test unit (computed): cm - ellipse(at top 50cm left 50%)
[PASS] test unit (computed): cm - ellipse(at top 50cm left 50cm)
[PASS] test unit (computed): cm - ellipse(at top 50cm right 80%)
[PASS] test unit (computed): cm - ellipse(at bottom 70% left 50cm)
[FAIL] test unit (computed): cm - ellipse(at bottom 70% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 30%)" but got "ellipse(at right 3023.62px top 30%)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 3023.62px bottom 2645.67px)"
[PASS] test unit (computed): cm - ellipse(at right 80% top 50cm)
[FAIL] test unit (computed): cm - ellipse(at right 80% bottom 70cm)
  assert_equals: expected "ellipse(at 20% calc(100% - 2645.67px))" but got "ellipse(at left 20% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 3023.62px top 50%)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 30%)" but got "ellipse(at right 3023.62px top 30%)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm bottom 70cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 3023.62px bottom 2645.67px)"
[PASS] test unit (computed): mm - ellipse(at 50mm)
[PASS] test unit (computed): mm - ellipse(at 50% 50mm)
[PASS] test unit (computed): mm - ellipse(at 50mm 50%)
[PASS] test unit (computed): mm - ellipse(at 50mm 50mm)
[PASS] test unit (computed): mm - ellipse(at left 50mm)
[PASS] test unit (computed): mm - ellipse(at 50mm top)
[PASS] test unit (computed): mm - ellipse(at right 80mm)
[PASS] test unit (computed): mm - ellipse(at 70mm bottom)
[PASS] test unit (computed): mm - ellipse(at center 60mm)
[PASS] test unit (computed): mm - ellipse(at 60mm center)
[PASS] test unit (computed): mm - ellipse(at left 50% top 50mm)
[FAIL] test unit (computed): mm - ellipse(at left 50% bottom 70mm)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[PASS] test unit (computed): mm - ellipse(at left 50mm top 50%)
[PASS] test unit (computed): mm - ellipse(at left 50mm top 50mm)
[PASS] test unit (computed): mm - ellipse(at left 50mm bottom 70%)
[PASS] test unit (computed): mm - ellipse(at top 50% left 50mm)
[FAIL] test unit (computed): mm - ellipse(at top 50% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 302.362px top 50%)"
[PASS] test unit (computed): mm - ellipse(at top 50mm left 50%)
[PASS] test unit (computed): mm - ellipse(at top 50mm left 50mm)
[PASS] test unit (computed): mm - ellipse(at top 50mm right 80%)
[PASS] test unit (computed): mm - ellipse(at bottom 70% left 50mm)
[FAIL] test unit (computed): mm - ellipse(at bottom 70% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 30%)" but got "ellipse(at right 302.362px top 30%)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 302.362px bottom 264.567px)"
[PASS] test unit (computed): mm - ellipse(at right 80% top 50mm)
[FAIL] test unit (computed): mm - ellipse(at right 80% bottom 70mm)
  assert_equals: expected "ellipse(at 20% calc(100% - 264.567px))" but got "ellipse(at left 20% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 302.362px top 50%)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 30%)" but got "ellipse(at right 302.362px top 30%)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm bottom 70mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 302.362px bottom 264.567px)"
[PASS] test unit (computed): in - ellipse(at 50in)
[PASS] test unit (computed): in - ellipse(at 50% 50in)
[PASS] test unit (computed): in - ellipse(at 50in 50%)
[PASS] test unit (computed): in - ellipse(at 50in 50in)
[PASS] test unit (computed): in - ellipse(at left 50in)
[PASS] test unit (computed): in - ellipse(at 50in top)
[PASS] test unit (computed): in - ellipse(at right 80in)
[PASS] test unit (computed): in - ellipse(at 70in bottom)
[PASS] test unit (computed): in - ellipse(at center 60in)
[PASS] test unit (computed): in - ellipse(at 60in center)
[PASS] test unit (computed): in - ellipse(at left 50% top 50in)
[FAIL] test unit (computed): in - ellipse(at left 50% bottom 70in)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[PASS] test unit (computed): in - ellipse(at left 50in top 50%)
[PASS] test unit (computed): in - ellipse(at left 50in top 50in)
[PASS] test unit (computed): in - ellipse(at left 50in bottom 70%)
[PASS] test unit (computed): in - ellipse(at top 50% left 50in)
[FAIL] test unit (computed): in - ellipse(at top 50% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 7680px top 50%)"
[PASS] test unit (computed): in - ellipse(at top 50in left 50%)
[PASS] test unit (computed): in - ellipse(at top 50in left 50in)
[PASS] test unit (computed): in - ellipse(at top 50in right 80%)
[PASS] test unit (computed): in - ellipse(at bottom 70% left 50in)
[FAIL] test unit (computed): in - ellipse(at bottom 70% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 30%)" but got "ellipse(at right 7680px top 30%)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 7680px bottom 6720px)"
[PASS] test unit (computed): in - ellipse(at right 80% top 50in)
[FAIL] test unit (computed): in - ellipse(at right 80% bottom 70in)
  assert_equals: expected "ellipse(at 20% calc(100% - 6720px))" but got "ellipse(at left 20% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at right 80in top 50%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 7680px top 50%)"
[FAIL] test unit (computed): in - ellipse(at right 80in bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 30%)" but got "ellipse(at right 7680px top 30%)"
[FAIL] test unit (computed): in - ellipse(at right 80in bottom 70in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 7680px bottom 6720px)"
[PASS] test unit (computed): pt - ellipse(at 50pt)
[PASS] test unit (computed): pt - ellipse(at 50% 50pt)
[PASS] test unit (computed): pt - ellipse(at 50pt 50%)
[PASS] test unit (computed): pt - ellipse(at 50pt 50pt)
[PASS] test unit (computed): pt - ellipse(at left 50pt)
[PASS] test unit (computed): pt - ellipse(at 50pt top)
[PASS] test unit (computed): pt - ellipse(at right 80pt)
[PASS] test unit (computed): pt - ellipse(at 70pt bottom)
[PASS] test unit (computed): pt - ellipse(at center 60pt)
[PASS] test unit (computed): pt - ellipse(at 60pt center)
[PASS] test unit (computed): pt - ellipse(at left 50% top 50pt)
[FAIL] test unit (computed): pt - ellipse(at left 50% bottom 70pt)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[PASS] test unit (computed): pt - ellipse(at left 50pt top 50%)
[PASS] test unit (computed): pt - ellipse(at left 50pt top 50pt)
[PASS] test unit (computed): pt - ellipse(at left 50pt bottom 70%)
[PASS] test unit (computed): pt - ellipse(at top 50% left 50pt)
[FAIL] test unit (computed): pt - ellipse(at top 50% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 106.667px top 50%)"
[PASS] test unit (computed): pt - ellipse(at top 50pt left 50%)
[PASS] test unit (computed): pt - ellipse(at top 50pt left 50pt)
[PASS] test unit (computed): pt - ellipse(at top 50pt right 80%)
[PASS] test unit (computed): pt - ellipse(at bottom 70% left 50pt)
[FAIL] test unit (computed): pt - ellipse(at bottom 70% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 30%)" but got "ellipse(at right 106.667px top 30%)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 106.667px bottom 93.3333px)"
[PASS] test unit (computed): pt - ellipse(at right 80% top 50pt)
[FAIL] test unit (computed): pt - ellipse(at right 80% bottom 70pt)
  assert_equals: expected "ellipse(at 20% calc(100% - 93.3333px))" but got "ellipse(at left 20% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt top 50%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 106.667px top 50%)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 30%)" but got "ellipse(at right 106.667px top 30%)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt bottom 70pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 106.667px bottom 93.3333px)"
[PASS] test unit (computed): pc - ellipse(at 50pc)
[PASS] test unit (computed): pc - ellipse(at 50% 50pc)
[PASS] test unit (computed): pc - ellipse(at 50pc 50%)
[PASS] test unit (computed): pc - ellipse(at 50pc 50pc)
[PASS] test unit (computed): pc - ellipse(at left 50pc)
[PASS] test unit (computed): pc - ellipse(at 50pc top)
[PASS] test unit (computed): pc - ellipse(at right 80pc)
[PASS] test unit (computed): pc - ellipse(at 70pc bottom)
[PASS] test unit (computed): pc - ellipse(at center 60pc)
[PASS] test unit (computed): pc - ellipse(at 60pc center)
[PASS] test unit (computed): pc - ellipse(at left 50% top 50pc)
[FAIL] test unit (computed): pc - ellipse(at left 50% bottom 70pc)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[PASS] test unit (computed): pc - ellipse(at left 50pc top 50%)
[PASS] test unit (computed): pc - ellipse(at left 50pc top 50pc)
[PASS] test unit (computed): pc - ellipse(at left 50pc bottom 70%)
[PASS] test unit (computed): pc - ellipse(at top 50% left 50pc)
[FAIL] test unit (computed): pc - ellipse(at top 50% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[PASS] test unit (computed): pc - ellipse(at top 50pc left 50%)
[PASS] test unit (computed): pc - ellipse(at top 50pc left 50pc)
[PASS] test unit (computed): pc - ellipse(at top 50pc right 80%)
[PASS] test unit (computed): pc - ellipse(at bottom 70% left 50pc)
[FAIL] test unit (computed): pc - ellipse(at bottom 70% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): pc - ellipse(at right 80% top 50pc)
[FAIL] test unit (computed): pc - ellipse(at right 80% bottom 70pc)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc bottom 70pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): em - ellipse(at 50em)
[PASS] test unit (computed): em - ellipse(at 50% 50em)
[PASS] test unit (computed): em - ellipse(at 50em 50%)
[PASS] test unit (computed): em - ellipse(at 50em 50em)
[PASS] test unit (computed): em - ellipse(at left 50em)
[PASS] test unit (computed): em - ellipse(at 50em top)
[PASS] test unit (computed): em - ellipse(at right 80em)
[PASS] test unit (computed): em - ellipse(at 70em bottom)
[PASS] test unit (computed): em - ellipse(at center 60em)
[PASS] test unit (computed): em - ellipse(at 60em center)
[PASS] test unit (computed): em - ellipse(at left 50% top 50em)
[FAIL] test unit (computed): em - ellipse(at left 50% bottom 70em)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[PASS] test unit (computed): em - ellipse(at left 50em top 50%)
[PASS] test unit (computed): em - ellipse(at left 50em top 50em)
[PASS] test unit (computed): em - ellipse(at left 50em bottom 70%)
[PASS] test unit (computed): em - ellipse(at top 50% left 50em)
[FAIL] test unit (computed): em - ellipse(at top 50% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[PASS] test unit (computed): em - ellipse(at top 50em left 50%)
[PASS] test unit (computed): em - ellipse(at top 50em left 50em)
[PASS] test unit (computed): em - ellipse(at top 50em right 80%)
[PASS] test unit (computed): em - ellipse(at bottom 70% left 50em)
[FAIL] test unit (computed): em - ellipse(at bottom 70% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): em - ellipse(at right 80% top 50em)
[FAIL] test unit (computed): em - ellipse(at right 80% bottom 70em)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at right 80em top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): em - ellipse(at right 80em bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): em - ellipse(at right 80em bottom 70em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): ex - ellipse(at 50ex)
[PASS] test unit (computed): ex - ellipse(at 50% 50ex)
[PASS] test unit (computed): ex - ellipse(at 50ex 50%)
[PASS] test unit (computed): ex - ellipse(at 50ex 50ex)
[PASS] test unit (computed): ex - ellipse(at left 50ex)
[PASS] test unit (computed): ex - ellipse(at 50ex top)
[PASS] test unit (computed): ex - ellipse(at right 80ex)
[PASS] test unit (computed): ex - ellipse(at 70ex bottom)
[PASS] test unit (computed): ex - ellipse(at center 60ex)
[PASS] test unit (computed): ex - ellipse(at 60ex center)
[PASS] test unit (computed): ex - ellipse(at left 50% top 50ex)
[FAIL] test unit (computed): ex - ellipse(at left 50% bottom 70ex)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[PASS] test unit (computed): ex - ellipse(at left 50ex top 50%)
[PASS] test unit (computed): ex - ellipse(at left 50ex top 50ex)
[PASS] test unit (computed): ex - ellipse(at left 50ex bottom 70%)
[PASS] test unit (computed): ex - ellipse(at top 50% left 50ex)
[FAIL] test unit (computed): ex - ellipse(at top 50% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 50%)" but got "ellipse(at right 1024px top 50%)"
[PASS] test unit (computed): ex - ellipse(at top 50ex left 50%)
[PASS] test unit (computed): ex - ellipse(at top 50ex left 50ex)
[PASS] test unit (computed): ex - ellipse(at top 50ex right 80%)
[PASS] test unit (computed): ex - ellipse(at bottom 70% left 50ex)
[FAIL] test unit (computed): ex - ellipse(at bottom 70% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 30%)" but got "ellipse(at right 1024px top 30%)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) calc(100% - 896px))" but got "ellipse(at right 1024px bottom 896px)"
[PASS] test unit (computed): ex - ellipse(at right 80% top 50ex)
[FAIL] test unit (computed): ex - ellipse(at right 80% bottom 70ex)
  assert_equals: expected "ellipse(at 20% calc(100% - 896px))" but got "ellipse(at left 20% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 50%)" but got "ellipse(at right 1024px top 50%)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 30%)" but got "ellipse(at right 1024px top 30%)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex bottom 70ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) calc(100% - 896px))" but got "ellipse(at right 1024px bottom 896px)"
[PASS] test unit (computed): ch - ellipse(at 50ch)
[PASS] test unit (computed): ch - ellipse(at 50% 50ch)
[PASS] test unit (computed): ch - ellipse(at 50ch 50%)
[PASS] test unit (computed): ch - ellipse(at 50ch 50ch)
[PASS] test unit (computed): ch - ellipse(at left 50ch)
[PASS] test unit (computed): ch - ellipse(at 50ch top)
[PASS] test unit (computed): ch - ellipse(at right 80ch)
[PASS] test unit (computed): ch - ellipse(at 70ch bottom)
[PASS] test unit (computed): ch - ellipse(at center 60ch)
[PASS] test unit (computed): ch - ellipse(at 60ch center)
[PASS] test unit (computed): ch - ellipse(at left 50% top 50ch)
[FAIL] test unit (computed): ch - ellipse(at left 50% bottom 70ch)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[PASS] test unit (computed): ch - ellipse(at left 50ch top 50%)
[PASS] test unit (computed): ch - ellipse(at left 50ch top 50ch)
[PASS] test unit (computed): ch - ellipse(at left 50ch bottom 70%)
[PASS] test unit (computed): ch - ellipse(at top 50% left 50ch)
[FAIL] test unit (computed): ch - ellipse(at top 50% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[PASS] test unit (computed): ch - ellipse(at top 50ch left 50%)
[PASS] test unit (computed): ch - ellipse(at top 50ch left 50ch)
[PASS] test unit (computed): ch - ellipse(at top 50ch right 80%)
[PASS] test unit (computed): ch - ellipse(at bottom 70% left 50ch)
[FAIL] test unit (computed): ch - ellipse(at bottom 70% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): ch - ellipse(at right 80% top 50ch)
[FAIL] test unit (computed): ch - ellipse(at right 80% bottom 70ch)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch bottom 70ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): rem - ellipse(at 50rem)
[PASS] test unit (computed): rem - ellipse(at 50% 50rem)
[PASS] test unit (computed): rem - ellipse(at 50rem 50%)
[PASS] test unit (computed): rem - ellipse(at 50rem 50rem)
[PASS] test unit (computed): rem - ellipse(at left 50rem)
[PASS] test unit (computed): rem - ellipse(at 50rem top)
[PASS] test unit (computed): rem - ellipse(at right 80rem)
[PASS] test unit (computed): rem - ellipse(at 70rem bottom)
[PASS] test unit (computed): rem - ellipse(at center 60rem)
[PASS] test unit (computed): rem - ellipse(at 60rem center)
[PASS] test unit (computed): rem - ellipse(at left 50% top 50rem)
[FAIL] test unit (computed): rem - ellipse(at left 50% bottom 70rem)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[PASS] test unit (computed): rem - ellipse(at left 50rem top 50%)
[PASS] test unit (computed): rem - ellipse(at left 50rem top 50rem)
[PASS] test unit (computed): rem - ellipse(at left 50rem bottom 70%)
[PASS] test unit (computed): rem - ellipse(at top 50% left 50rem)
[FAIL] test unit (computed): rem - ellipse(at top 50% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[PASS] test unit (computed): rem - ellipse(at top 50rem left 50%)
[PASS] test unit (computed): rem - ellipse(at top 50rem left 50rem)
[PASS] test unit (computed): rem - ellipse(at top 50rem right 80%)
[PASS] test unit (computed): rem - ellipse(at bottom 70% left 50rem)
[FAIL] test unit (computed): rem - ellipse(at bottom 70% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): rem - ellipse(at right 80% top 50rem)
[FAIL] test unit (computed): rem - ellipse(at right 80% bottom 70rem)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem bottom 70rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[PASS] test unit (computed): vw - ellipse(at 50vw)
[PASS] test unit (computed): vw - ellipse(at 50% 50vw)
[PASS] test unit (computed): vw - ellipse(at 50vw 50%)
[PASS] test unit (computed): vw - ellipse(at 50vw 50vw)
[PASS] test unit (computed): vw - ellipse(at left 50vw)
[PASS] test unit (computed): vw - ellipse(at 50vw top)
[PASS] test unit (computed): vw - ellipse(at right 80vw)
[PASS] test unit (computed): vw - ellipse(at 70vw bottom)
[PASS] test unit (computed): vw - ellipse(at center 60vw)
[PASS] test unit (computed): vw - ellipse(at 60vw center)
[PASS] test unit (computed): vw - ellipse(at left 50% top 50vw)
[FAIL] test unit (computed): vw - ellipse(at left 50% bottom 70vw)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[PASS] test unit (computed): vw - ellipse(at left 50vw top 50%)
[PASS] test unit (computed): vw - ellipse(at left 50vw top 50vw)
[PASS] test unit (computed): vw - ellipse(at left 50vw bottom 70%)
[PASS] test unit (computed): vw - ellipse(at top 50% left 50vw)
[FAIL] test unit (computed): vw - ellipse(at top 50% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[PASS] test unit (computed): vw - ellipse(at top 50vw left 50%)
[PASS] test unit (computed): vw - ellipse(at top 50vw left 50vw)
[PASS] test unit (computed): vw - ellipse(at top 50vw right 80%)
[PASS] test unit (computed): vw - ellipse(at bottom 70% left 50vw)
[FAIL] test unit (computed): vw - ellipse(at bottom 70% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[PASS] test unit (computed): vw - ellipse(at right 80% top 50vw)
[FAIL] test unit (computed): vw - ellipse(at right 80% bottom 70vw)
  assert_equals: expected "ellipse(at 20% calc(100% - 560px))" but got "ellipse(at left 20% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw top 50%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw bottom 70vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[PASS] test unit (computed): vh - ellipse(at 50vh)
[PASS] test unit (computed): vh - ellipse(at 50% 50vh)
[PASS] test unit (computed): vh - ellipse(at 50vh 50%)
[PASS] test unit (computed): vh - ellipse(at 50vh 50vh)
[PASS] test unit (computed): vh - ellipse(at left 50vh)
[PASS] test unit (computed): vh - ellipse(at 50vh top)
[PASS] test unit (computed): vh - ellipse(at right 80vh)
[PASS] test unit (computed): vh - ellipse(at 70vh bottom)
[PASS] test unit (computed): vh - ellipse(at center 60vh)
[PASS] test unit (computed): vh - ellipse(at 60vh center)
[PASS] test unit (computed): vh - ellipse(at left 50% top 50vh)
[FAIL] test unit (computed): vh - ellipse(at left 50% bottom 70vh)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[PASS] test unit (computed): vh - ellipse(at left 50vh top 50%)
[PASS] test unit (computed): vh - ellipse(at left 50vh top 50vh)
[PASS] test unit (computed): vh - ellipse(at left 50vh bottom 70%)
[PASS] test unit (computed): vh - ellipse(at top 50% left 50vh)
[FAIL] test unit (computed): vh - ellipse(at top 50% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[PASS] test unit (computed): vh - ellipse(at top 50vh left 50%)
[PASS] test unit (computed): vh - ellipse(at top 50vh left 50vh)
[PASS] test unit (computed): vh - ellipse(at top 50vh right 80%)
[PASS] test unit (computed): vh - ellipse(at bottom 70% left 50vh)
[FAIL] test unit (computed): vh - ellipse(at bottom 70% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[PASS] test unit (computed): vh - ellipse(at right 80% top 50vh)
[FAIL] test unit (computed): vh - ellipse(at right 80% bottom 70vh)
  assert_equals: expected "ellipse(at 20% calc(100% - 420px))" but got "ellipse(at left 20% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh top 50%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh bottom 70vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[PASS] test unit (computed): vmin - ellipse(at 50vmin)
[PASS] test unit (computed): vmin - ellipse(at 50% 50vmin)
[PASS] test unit (computed): vmin - ellipse(at 50vmin 50%)
[PASS] test unit (computed): vmin - ellipse(at 50vmin 50vmin)
[PASS] test unit (computed): vmin - ellipse(at left 50vmin)
[PASS] test unit (computed): vmin - ellipse(at 50vmin top)
[PASS] test unit (computed): vmin - ellipse(at right 80vmin)
[PASS] test unit (computed): vmin - ellipse(at 70vmin bottom)
[PASS] test unit (computed): vmin - ellipse(at center 60vmin)
[PASS] test unit (computed): vmin - ellipse(at 60vmin center)
[PASS] test unit (computed): vmin - ellipse(at left 50% top 50vmin)
[FAIL] test unit (computed): vmin - ellipse(at left 50% bottom 70vmin)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[PASS] test unit (computed): vmin - ellipse(at left 50vmin top 50%)
[PASS] test unit (computed): vmin - ellipse(at left 50vmin top 50vmin)
[PASS] test unit (computed): vmin - ellipse(at left 50vmin bottom 70%)
[PASS] test unit (computed): vmin - ellipse(at top 50% left 50vmin)
[FAIL] test unit (computed): vmin - ellipse(at top 50% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[PASS] test unit (computed): vmin - ellipse(at top 50vmin left 50%)
[PASS] test unit (computed): vmin - ellipse(at top 50vmin left 50vmin)
[PASS] test unit (computed): vmin - ellipse(at top 50vmin right 80%)
[PASS] test unit (computed): vmin - ellipse(at bottom 70% left 50vmin)
[FAIL] test unit (computed): vmin - ellipse(at bottom 70% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[PASS] test unit (computed): vmin - ellipse(at right 80% top 50vmin)
[FAIL] test unit (computed): vmin - ellipse(at right 80% bottom 70vmin)
  assert_equals: expected "ellipse(at 20% calc(100% - 420px))" but got "ellipse(at left 20% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin top 50%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin bottom 70vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[PASS] test unit (computed): vmax - ellipse(at 50vmax)
[PASS] test unit (computed): vmax - ellipse(at 50% 50vmax)
[PASS] test unit (computed): vmax - ellipse(at 50vmax 50%)
[PASS] test unit (computed): vmax - ellipse(at 50vmax 50vmax)
[PASS] test unit (computed): vmax - ellipse(at left 50vmax)
[PASS] test unit (computed): vmax - ellipse(at 50vmax top)
[PASS] test unit (computed): vmax - ellipse(at right 80vmax)
[PASS] test unit (computed): vmax - ellipse(at 70vmax bottom)
[PASS] test unit (computed): vmax - ellipse(at center 60vmax)
[PASS] test unit (computed): vmax - ellipse(at 60vmax center)
[PASS] test unit (computed): vmax - ellipse(at left 50% top 50vmax)
[FAIL] test unit (computed): vmax - ellipse(at left 50% bottom 70vmax)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[PASS] test unit (computed): vmax - ellipse(at left 50vmax top 50%)
[PASS] test unit (computed): vmax - ellipse(at left 50vmax top 50vmax)
[PASS] test unit (computed): vmax - ellipse(at left 50vmax bottom 70%)
[PASS] test unit (computed): vmax - ellipse(at top 50% left 50vmax)
[FAIL] test unit (computed): vmax - ellipse(at top 50% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[PASS] test unit (computed): vmax - ellipse(at top 50vmax left 50%)
[PASS] test unit (computed): vmax - ellipse(at top 50vmax left 50vmax)
[PASS] test unit (computed): vmax - ellipse(at top 50vmax right 80%)
[PASS] test unit (computed): vmax - ellipse(at bottom 70% left 50vmax)
[FAIL] test unit (computed): vmax - ellipse(at bottom 70% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[PASS] test unit (computed): vmax - ellipse(at right 80% top 50vmax)
[FAIL] test unit (computed): vmax - ellipse(at right 80% bottom 70vmax)
  assert_equals: expected "ellipse(at 20% calc(100% - 560px))" but got "ellipse(at left 20% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax top 50%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax bottom 70vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
Harness: the test ran to completion.
